refactor: 좌표 입력 파싱과 검증의 책임을 view로 변경
