+++ src/main/java/controller/JanggiController.java
@@ -2,13 +2,11 @@ package controller;
 
 import domain.BoardPosition;
 import domain.Janggi;
+import java.util.List;
 import view.InputView;
 import view.OutputView;
 
 public class JanggiController {
-
-    private static final int COORDINATE_PARTS_COUNT = 2;
-    private static final String POSITION_DELIMITER = ",";
 
     private final InputView inputView;
     private final OutputView outputView;
@@ -26,45 +24,25 @@ public class JanggiController {
         while (true) {
             outputView.printBoard(janggi.getPieces(), janggi.getCurrentTeam());
 
-            final String selectPosition = inputView.inputSelectPosition();
+            final List<Integer> selectPosition = inputView.inputSelectPosition();
             final BoardPosition selectBoardPosition = createBoardPosition(selectPosition);
             janggi.validateSelectedPiece(selectBoardPosition);
-            
-            final String destinationPosition = inputView.inputDestinationPosition();
+
+            final List<Integer> destinationPosition = inputView.inputDestinationPosition();
             final BoardPosition destinationBoardPosition = createBoardPosition(destinationPosition);
 
             janggi.processTurn(selectBoardPosition, destinationBoardPosition);
         }
     }
 
-    public BoardPosition createBoardPosition(final String inputPosition) {
-        final int[] coordinates = validateAndParse(validateAndSplit(inputPosition));
-        return new BoardPosition(coordinates[0], coordinates[1]);
+    public BoardPosition createBoardPosition(final List<Integer> coordinates) {
+        validateSize(coordinates);
+        return new BoardPosition(coordinates.get(0), coordinates.get(1));
     }
 
-    private String[] validateAndSplit(final String input) {
-        final String[] coordinates;
-        try {
-            coordinates = input.split(POSITION_DELIMITER, -1);
-        } catch (Exception e) {
-            throw new IllegalArgumentException("좌표는 'x,y' 형식으로 입력해야 합니다.");
-        }
-
-        if (coordinates.length != COORDINATE_PARTS_COUNT) {
-            throw new IllegalArgumentException("좌표는 'x,y' 형식으로 입력해야 합니다.");
-        }
-
-        return coordinates;
-    }
-
-    private int[] validateAndParse(final String[] parts) {
-        try {
-            final int x = Integer.parseInt(parts[0].trim());
-            final int y = Integer.parseInt(parts[1].trim());
-
-            return new int[]{x, y};
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("좌표는 숫자로 입력해야 합니다.");
+    private void validateSize(final List<Integer> coordinates) {
+        if (coordinates.size() != BoardPosition.COORDINATE_PARTS_COUNT) {
+            throw new IllegalArgumentException("좌표의 크기는 2여야 합니다.");
         }
     }
 }
